fix: memory and speed improvement
fix: memory and speed improvement

I originally was being extremely inefficient in loading images in this model by first reading in all the images to a list, then transforming to an array, and finally performing image re-sizing onto the array. This caused much higher memory to be used than needed, as well as vastly slowing things down. Now, the images are re-sized as desired right after being read and before being appended to the list (and made into an array).